refactor: include stories for typescript compiler
