+++ src/components/ebay-switch/switch.stories.ts
@@ -1,12 +1,13 @@
+import type { Story } from "@storybook/marko";
 import { tagToString } from "../../../.storybook/storybook-code-source";
 import Readme from "./README.md";
-import Component from "./index.marko";
+import Component, { type Input } from "./index.marko";
 import WithLabelTemplate from "./examples/with-label.marko";
 import DisabledTemplate from "./examples/disabled-with-label.marko";
 import WithLabelCode from "./examples/with-label.marko?raw";
 import DisabledCode from "./examples/disabled-with-label.marko?raw";
 
-const Template = (args) => ({ input: args });
+const Template: Story<Input> = (args) => ({ input: args });
 
 export default {
     title: "form input/ebay-switch",
@@ -33,7 +34,7 @@ export default {
     },
 };
 
-export const WithLabel = (args) => ({
+export const WithLabel: Story<Input> = (args) => ({
     input: args,
     component: WithLabelTemplate,
 });
@@ -46,7 +47,7 @@ WithLabel.parameters = {
     },
 };
 
-export const Disabled = (args) => ({
+export const Disabled: Story<Input> = (args) => ({
     input: args,
     component: DisabledTemplate,
 });
