--- src/global.d.ts
@@ -1,4 +1,0 @@
-declare module "*.marko?raw" {
-    const value: string;
-    export default value;
-}
